+++ node/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   admission.
 - Reject blob transactions whose gas charge overflows during mempool
   admission
+- OutOfSync: stop consensus task on entry
 
 ## [1.6.0] - 2026-02-27